casting bug

+++ src/neural_eval.rs
@@ -86,7 +86,7 @@ fn serialize(position: &Chess) -> [i16; L0] {
         Color::Black => 0,
         Color::White => 1,
     };
-    result[769] = position.fullmoves() as i16;
+    result[769] = u32::from(position.fullmoves()) as i16;
 
     return result;
 }
